refactor(lib): move askUser* cluster into src/lib/askUser/
First concrete slice of src/lib domain organization: the 8-file
ask_user questionnaire domain (payload validation, form model, keyboard,
options layout, settle/timeout/clocks, demo path) + its tests + the
askUserBar guard test move into src/lib/askUser/, matching the existing
pet/ session/ remoteIm/ subdirectory convention (prefix kept in names).

Only 9 external importers — all updated to @/lib/askUser/*; internal
./session, ./grokCatalog, ./gateClock references became ../. The guard
test's ../app path was rebased to ../../app.
